:recycle: Move getSignaturesBuffer to utils
